Simplify proposal package
In this commit I removed a lot of legacy directives.

The idea of having separate directives for sheets may be beneficial in
the long term, but for now it is confusing at best. So I inlined their
templates and removed the directives.

Since most of this code will be rewritten in this branch, minor
functionality setbacks are affordable.